add test helper
