+++ tests/test_users_api.py
@@ -1,75 +1,10 @@
-from pytest import fixture
-from app.models import User
-from app import create_app, db
-from base64 import b64encode
-from flask_jwt_extended import create_access_token
-
-
-def generate_users():
-    u1 = User(
-        username='test',
-        password='pass12A@',
-        admin_role=False
-    )
-
-    u2 = User(
-        username='admin',
-        password='pass12A@',
-        admin_role=True
-    )
-
-    u3 = User(
-        username='edit',
-        password='pass12A@',
-        admin_role=False
-    )
-
-    db.session.add(u1)
-    db.session.add(u2)
-    db.session.add(u3)
-    
-    db.session.commit()
-
-
-@fixture
-def application():
-    # set up
-    app = create_app('testing')
-    app_context = app.app_context()
-    app_context.push()
-    db.create_all()
-
-    generate_users()
-
-    yield app
-    
-    # teardown
-    db.session.remove()
-    db.drop_all()
-    app_context.pop()
-
-
-def _get_jwt(test_client, username):
-    user = User.query.filter_by(username=username).first()
-    
-    credentials = b64encode(f"{user.username}:pass12A@".encode('utf-8')).decode('utf-8')
-
-    response = test_client.get(
-        '/api/v1/login',
-        headers={"Authorization": f"Basic {credentials}"}
-    )
-
-    assert response.status_code == 200
-    assert response.get_json().get('token', False)
-    token = response.get_json()['token']
-    
-    return token
+from .helper import get_jwt, application
 
 
 def test_get_all_users(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'test')
+    token = get_jwt(test_client, 'test')
     response = test_client.get(
         '/api/v1/user',
         headers={'Authorization': f'Bearer {token}'}
@@ -87,20 +22,24 @@ def test_get_all_users_without_jwt(application):
     )
 
     assert response.status_code == 401
-    assert response.get_json() == {'msg': 'Missing Authorization Header'}
+
+    msg = response.get_json().get('msg', False)
+    assert msg == 'Missing Authorization Header'
 
 
 def test_get_user(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'test')
+    token = get_jwt(test_client, 'test')
     response = test_client.get(
         '/api/v1/user/1',
         headers={'Authorization': f'Bearer {token}'}
     )
 
     assert response.status_code == 200
-    assert response.get_json().get('user', False)
+
+    data = response.get_json().get('user', False)
+    assert isinstance(data, dict)
 
 
 def test_get_user_without_jwt(application):
@@ -111,13 +50,15 @@ def test_get_user_without_jwt(application):
     )
 
     assert response.status_code == 401
-    assert response.get_json() == {'msg': 'Missing Authorization Header'}
+
+    msg = response.get_json().get('msg', False)
+    assert msg == 'Missing Authorization Header'
 
 
 def test_get_user_not_found_error(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'test')
+    token = get_jwt(test_client, 'test')
     response = test_client.get(
         '/api/v1/user/100',
         headers={'Authorization': f'Bearer {token}'}
@@ -130,7 +71,7 @@ def test_get_user_not_found_error(application):
 def test_create_user(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'admin')
+    token = get_jwt(test_client, 'admin')
     response = test_client.post(
         '/api/v1/user',
         headers={'Authorization': f'Bearer {token}'},
@@ -146,7 +87,7 @@ def test_create_user(application):
 def test_create_user_permission_error(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'test')
+    token = get_jwt(test_client, 'test')
     response = test_client.post(
         '/api/v1/user',
         headers={'Authorization': f'Bearer {token}'},
@@ -162,7 +103,7 @@ def test_create_user_permission_error(application):
 def test_create_user_mimetype_error(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'test')
+    token = get_jwt(test_client, 'test')
     response = test_client.post(
         '/api/v1/user',
         headers={'Authorization': f'Bearer {token}'},
@@ -178,7 +119,7 @@ def test_create_user_mimetype_error(application):
 def test_create_user_not_valid_data(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'admin')
+    token = get_jwt(test_client, 'admin')
     response = test_client.post(
         '/api/v1/user',
         headers={'Authorization': f'Bearer {token}'},
@@ -194,7 +135,7 @@ def test_create_user_not_valid_data(application):
 def test_edit_user(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'admin')
+    token = get_jwt(test_client, 'admin')
     response = test_client.put(
         '/api/v1/user/3',
         headers={'Authorization': f'Bearer {token}'},
@@ -210,7 +151,7 @@ def test_edit_user(application):
 def test_edit_user_notfound_error(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'admin')
+    token = get_jwt(test_client, 'admin')
     response = test_client.put(
         '/api/v1/user/300',
         headers={'Authorization': f'Bearer {token}'},
@@ -226,7 +167,7 @@ def test_edit_user_notfound_error(application):
 def test_edit_user_permission_error(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'test')
+    token = get_jwt(test_client, 'test')
     response = test_client.put(
         '/api/v1/user/3',
         headers={'Authorization': f'Bearer {token}'},
@@ -257,7 +198,7 @@ def test_edit_user_without_jwt(application):
 def test_delete_user(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'admin')
+    token = get_jwt(test_client, 'admin')
     response = test_client.delete(
         '/api/v1/user/3',
         headers={'Authorization': f'Bearer {token}'}
@@ -272,7 +213,7 @@ def test_delete_user(application):
 def test_delete_user_notfound_error(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'admin')
+    token = get_jwt(test_client, 'admin')
     response = test_client.delete(
         '/api/v1/user/300',
         headers={'Authorization': f'Bearer {token}'}
@@ -287,7 +228,7 @@ def test_delete_user_notfound_error(application):
 def test_delete_user_permission_error(application):
     test_client = application.test_client()
 
-    token = _get_jwt(test_client, 'test')
+    token = get_jwt(test_client, 'test')
     response = test_client.delete(
         '/api/v1/user/3',
         headers={'Authorization': f'Bearer {token}'}
